update bi HistoricoMetricasMetricas

+++ backend/src/main/java/com/example/backend/bi/historicoMetricas/HistoricoMetricasRepository.java
@@ -2,5 +2,5 @@ package com.example.backend.bi.historicoMetricas;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HistoricoMetricasRepository extends JpaRepository<HistoricoMetricas, Integer> {
+public interface HistoricoMetricasRepository extends JpaRepository<HistoricoMetricas, Long> {
 }
